+++ CHANGELOG.md
@@ -11,7 +11,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Add `bsiLoremIpsum()` function for js
 * Add functions to group for better extension: `addOrReplaceElements()`, `removeElementsWithId()`
 * Adds HTML_EDITOR_CONFIG_ID to enable editor configs for template parts
+* Add `scopePrefill` for prefilling Dropzones
+* Add shorthand `cx.Dropzone('dropzoneId', [require('elem1'), require('elem2')])`
 * Update Twing from 5.x to 7.x
+* Add handlebars functions to twig
+* Add `withExtendAllDropzones()` to content-part and template-part
 
 
 ## [2.0.1] - 03.12.2025